pyproject.toml
[project]
name = "nrc_ewas"
description = ""
authors = [
    {name = "Philibert de Mercey", email = "pdemercey@masae-analytics.com"},
]
readme = "README.md"
version = "0.1.2"

# --------------- requirements ---------------
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "psycopg2-binary>=2.9.10",
    "django-split-settings>=1.3.2",
    "django-modeltranslation>=0.19.16",
    "celery>=5.4.0",
    "redis>=5.0.8",
    "django-celery-beat>=2.7.0",
    "django-celery-results>=2.5.1",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "psycopg2>=2.9.10",
    "pandas>=2.3.2",
    "geopandas>=1.1.1",
    "dtmapi>=0.1.5",
    "openai>=1.107.3",
    "slack-sdk>=3.36.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "django-vite>=3.1.0",
    "rasterstats>=0.20.0",
    "titiler.application>=0.26.0",
    "uvicorn>=0.38.0",
    "openpyxl>=3.1.5",
]

# --------------- build/versioning ---------------
[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.8",
    "spatialite>=0.0.3",
    "fiona>=1.10.1",
    "playwright>=1.55.0",
    "pytest-playwright>=0.7.1",
    "pytest-django>=4.11.1",
    "pytest-xdist>=3.8.0",
]
production = [
    "uwsgi>=2.0.30",
]
# --------------- ruff ---------------
[tool.ruff]
line-length = 180
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle
    "W",    # pycodestyle
    "F",    # pyflake
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "D",    # pydocstyle
]

ignore = [
    # "E501",  # line-too-long
    "B017",  # pytest.raises(Exception) should be considered evil
    "B028",  # warnings.warn called without an explicit stacklevel keyword argument
    "B904",  # check for raise statements in exception handlers that lack a from clause
    "D203",  # incorrect-blank-line-before-class
    "D212",  # multi-line-summary-first-line
    "D105",  # undocumented-magic-method

    # The following are conflicting rules that ruff recommends to avoid
    # See https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",  # tab-indentation
    "E111",  # indentation-with-invalid-multiple
    "E114",  # indentation-with-invalid-multiple-comment
    "E117",  # over-indented
    "D206",  # docstring-tab-indentation
    "D300",  # triple-single-quotes
    "Q000",  # bad-quotes-inline-string
    "Q001",  # bad-quotes-multiline-string
    "Q002",  # bad-quotes-docstring
    "Q003",  # avoidable-escaped-quote
    "COM812",  # missing-trailing-comma
    "COM819",  # prohibited-trailing-comma
    "ISC002",  # multi-line-implicit-string-concatenation
]

exclude = [
    "**/docs",
]

[tool.ruff.lint.per-file-ignores]
# Ignore unused imports (F401) in these files
"__init__.py" = ["F401", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # Options: "google", "numpy", or "pep257"
